Rename error definition file
